refactor(chat): extract shared @endo/chat-kit base package
Create @endo/chat-kit, the shared message-rendering primitives the chat
bodies and host shell all consume, and move the six leaf utils into it:
markdown-render, markdown-vnodes, value-vnodes, time-formatters, chime,
locator. They author vnodes with plain `preact` `h` (the established
space-package pattern — confinement is applied by the host's
renderConfined, not the component), so the package depends only on
preact + @endo markmdown/pass-style/daemon/harden.

This is the prerequisite for the chat/channel package split: a space
package (space-chat next) can now get these utils from @endo/chat-kit
instead of importing back into the @endo/chat app, which would be
circular. No space package depends on @endo/chat, and this keeps it that
way.

All eleven in-app importers plus three test files are repointed from
`./x.js` to `@endo/chat-kit/x.js` (a pure path swap via the package's
subpath exports). chime's now-unused `@ts-expect-error` is replaced with
a cast. Build, lint, and the markdown/value/inbox/channel test suites
pass.
