+++ packages/chat-kit/package.json
@@ -1,0 +1,57 @@
+{
+  "name": "@endo/chat-kit",
+  "version": "0.1.0",
+  "private": true,
+  "description": "Shared message-rendering primitives for Endo Chat spaces (markdown/value vnodes, time formatting, chime, locator helpers)",
+  "keywords": [
+    "endo",
+    "chat",
+    "markdown",
+    "preact"
+  ],
+  "author": "Endo contributors",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/endojs/endo.git",
+    "directory": "packages/chat-kit"
+  },
+  "bugs": {
+    "url": "https://github.com/endojs/endo/issues"
+  },
+  "type": "module",
+  "exports": {
+    ".": "./index.js",
+    "./markdown-render.js": "./src/markdown-render.js",
+    "./markdown-vnodes.js": "./src/markdown-vnodes.js",
+    "./value-vnodes.js": "./src/value-vnodes.js",
+    "./time-formatters.js": "./src/time-formatters.js",
+    "./chime.js": "./src/chime.js",
+    "./locator.js": "./src/locator.js",
+    "./package.json": "./package.json"
+  },
+  "scripts": {
+    "build": "exit 0",
+    "lint": "yarn lint:eslint",
+    "lint-fix": "yarn lint:eslint --fix",
+    "lint:eslint": "eslint '**/*.js'"
+  },
+  "dependencies": {
+    "@endo/daemon": "workspace:^",
+    "@endo/harden": "workspace:^",
+    "@endo/markmdown": "workspace:^",
+    "@endo/pass-style": "workspace:^",
+    "preact": "^10.29.2"
+  },
+  "devDependencies": {
+    "eslint": "catalog:dev"
+  },
+  "eslintConfig": {
+    "extends": [
+      "plugin:@endo/internal"
+    ],
+    "env": {
+      "browser": true
+    }
+  }
+}
