+++ packages/chat-kit/src/markdown-render.js
