+++ packages/chat-kit/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../../tsconfig.eslint-base.json",
+  "include": ["*.js", "*.ts", "src"]
+}
