+++ packages/chat/forum-component.js
@@ -14,9 +14,9 @@ import {
   computeNodeContent,
   isEffectivelyDeleted,
 } from './edit-queue.js';
-import { relativeTime, timeFormatter } from './time-formatters.js';
-import { prepareTextWithPlaceholders } from './markdown-render.js';
-import { markdownToVnodes } from './markdown-vnodes.js';
+import { relativeTime, timeFormatter } from '@endo/chat-kit/time-formatters.js';
+import { prepareTextWithPlaceholders } from '@endo/chat-kit/markdown-render.js';
+import { markdownToVnodes } from '@endo/chat-kit/markdown-vnodes.js';
 
 /** @import { ChannelMessage } from './channel-utils.js' */
